requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
appnope=0.1.4=pypi_0
asttokens=2.4.1=pypi_0
bzip2=1.0.8=h6c40b1e_6
ca-certificates=2024.3.11=hecd8cb5_0
comm=0.2.2=pypi_0
debugpy=1.8.1=pypi_0
decorator=5.1.1=pypi_0
exceptiongroup=1.2.1=pypi_0
executing=2.0.1=pypi_0
ipykernel=6.29.4=pypi_0
ipython=8.25.0=pypi_0
jedi=0.19.1=pypi_0
jupyter-client=8.6.2=pypi_0
jupyter-core=5.7.2=pypi_0
libffi=3.4.4=hecd8cb5_1
matplotlib-inline=0.1.7=pypi_0
ncurses=6.4=hcec6c5f_0
nest-asyncio=1.6.0=pypi_0
numpy=1.23.5=pypi_0
opencv-python=4.10.0.82=pypi_0
openssl=3.0.14=h46256e1_0
packaging=24.1=pypi_0
parso=0.8.4=pypi_0
pexpect=4.9.0=pypi_0
pip=24.0=py310hecd8cb5_0
platformdirs=4.2.2=pypi_0
prompt-toolkit=3.0.47=pypi_0
psutil=5.9.8=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pygments=2.18.0=pypi_0
python=3.10.14=h5ee71fb_1
python-dateutil=2.9.0.post0=pypi_0
pyzmq=26.0.3=pypi_0
readline=8.2=hca72f7f_0
setuptools=70.0.0=pypi_0
six=1.16.0=pypi_0
sqlite=3.45.3=h6c40b1e_0
stack-data=0.6.3=pypi_0
tk=8.6.14=h4d00af3_0
tornado=6.4.1=pypi_0
traitlets=5.14.3=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024a=h04d1e81_0
wcwidth=0.2.13=pypi_0
wheel=0.43.0=py310hecd8cb5_0
xz=5.4.6=h6c40b1e_1
zlib=1.2.13=h4b97444_1
